feat(aws): Adapt detect secrets to this check
