+++ prowler/providers/aws/services/codebuild/codebuild_project_no_secrets_in_variables/codebuild_project_no_secrets_in_variables.metadata.json
@@ -1,7 +1,7 @@
 {
   "Provider": "aws",
-  "CheckID": "codebuild_project_no_plaintext_credentials",
-  "CheckTitle": "Ensure CodeBuild projects do not use plaintext AWS credentials",
+  "CheckID": "codebuild_project_no_secrets_in_variables",
+  "CheckTitle": "Ensure CodeBuild projects do not contain plaintext AWS credentials",
   "CheckType": [
     "Security Best Practices"
   ],
@@ -15,7 +15,7 @@
   "RelatedUrl": "https://docs.aws.amazon.com/codebuild/latest/userguide/how-to-create-project-console.html",
   "Remediation": {
     "Code": {
-      "CLI": "aws codebuild update-project --name {project-name} --source '{\"type\": \"NO_SOURCE\", \"buildspec\": \"{buildspec-content}\"}' --environment '{\"environmentVariables\": [{\"name\": \"VAR_NAME\", \"value\": \"VAR_VALUE\", \"type\": \"PARAMETER_STORE\"}]}'",
+      "CLI": "",
       "NativeIaC": "",
       "Other": "https://docs.aws.amazon.com/securityhub/latest/userguide/codebuild-controls.html",
       "Terraform": ""
